apps/core/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy
psycopg2-binary
redis
openai
google-generativeai
langchain
langchain-community
python-socketio
chromadb
pymupdf
sentence-transformers
python-multipart
PyJWT>=2.8.0
psutil>=5.9.0
